+++ webpack/app/components/Constants.js
@@ -1,3 +1,17 @@
 export const ItemTypes = {
   ROW: "row"
 }
+
+export const CONFIRM_TRANSITION_IDS = [
+    "cancel",
+    "close",
+    "deactivate",
+    "invalidate",
+    "reinstate",
+    "reject",
+    "remove",
+    "remove",
+    "retest",
+    "retract",
+    "unassign"
+]